GloVe2word/requirements.txt
Flask
gunicorn
numpy
sklearn
lark-parser
scipy